+++ src/components/headerAdmin/headerAdmin.js
@@ -1,5 +1,8 @@
+import Swal from 'sweetalert2';
+import { error, loading, success } from '../../helpers/alertas';
 import { confirmModal } from '../modales/modalConfirm';
 import headerHtml from  './headerAdmin.html?raw';
+import { post } from '../../helpers/api';
 
 
 export const renderHeaderAdmin = (elememto) => {
@@ -10,7 +13,7 @@ export const renderHeaderAdmin = (elememto) => {
 
 function cargarContenido() {
 
-    const nombreCompleto = `${localStorage.getItem('nombre')} ${localStorage.getItem('apellido')}`;
+    const nombreCompleto = `${localStorage.getItem('full_name')}`;
 
     let usernameHeader = document.querySelector('.profile__name');
     let usernameMenu = document.querySelector('.profile-menu__username');
@@ -18,15 +21,30 @@ function cargarContenido() {
     usernameMenu.textContent = nombreCompleto;
 }
 
-function logout() {
+async function logout() {
 
-    localStorage.clear();
-    window.location.href = '#/login';
+    loading('Cerrando sesión...');
+    
+    const response = await post([], 'logout');
+    
+    if(response.success) {
+        Swal.close();
+        success(response.message);
+        localStorage.clear();
+        window.location.href = '#/login';
+
+        return;
+    }
+
+    else {
+        Swal.close();
+        error(response.message);
+    }
 
 }
 
-document.addEventListener('click', (e) => {
+document.addEventListener('click', async (e) => {
 
-    if(e.target.closest('#logout')) logout();
+    if(e.target.closest('#logoutAdmin')) await logout();
 
 })
